+++ module_2/FuramaResort/src/model/furama_facility/Villa.java
@@ -7,7 +7,7 @@ public class Villa extends FuramaFacility {
 
 
 
-    public Villa(String id,String serviceName, int useableArea, int price, int maxNumOfPeo, String typeOfRent, String roomStandard, double poolArea, int numOfLevel) {
+    public Villa(String id, String serviceName, double useableArea, int price, int maxNumOfPeo, String typeOfRent, String roomStandard, double poolArea, int numOfLevel) {
         super(id,serviceName, useableArea, price, maxNumOfPeo, typeOfRent);
         this.roomStandard = roomStandard;
         this.poolArea = poolArea;
